requirements.txt
# Core Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Astronomy & FITS File Handling
astropy>=5.0

# Signal Processing & Time Series Analysis
pycwt>=0.3.0a22
PyEMD>=1.3.0

# Nonlinear Dynamics & Entropy
antropy>=0.1.4
nolds>=0.5.2

# Machine Learning
scikit-learn>=1.0
hdbscan>=0.8.27
umap-learn>=0.5.1

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
